# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "B7Store API",
        "description": "API completa para o sistema B7Store - E-commerce",
        "contact": {
            "email": "suporte@b7store.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        },
        {
            "url": "https://api.b7store.com",
            "description": "Servidor de Produção"
        }
    ],
    "paths": {
...
        "/api/cart/mount": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Montar carrinho",
                "description": "Retorna informações dos produtos baseado nos IDs fornecidos",
                "operationId": "86e6ea94047f883a425b7f56030af425",
                "parameters": [
                    {
                        "name": "ids",
                        "in": "query",
                        "description": "IDs dos produtos separados por vírgula",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1,2,3"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produtos do carrinho"
                    },
                    "500": {
                        "description": "Erro ao buscar produtos"
                    }
                }
            }
        },
...
        "/api/cart/finish": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Finalizar compra",
                "description": "Cria um pedido com os itens do carrinho",
                "operationId": "537f9c2ca1010e978c2355a21153367b",
                "responses": {
                    "200": {
                        "description": "Pedido criado"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "500": {
                        "description": "Erro ao criar pedido"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/category/{slug}/metadata": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Obter metadados de uma categoria",
                "description": "Retorna informações da categoria e seus metadados com valores",
                "operationId": "a686b40c73dbbec054f5f6c4df6f6b50",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug da categoria",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "smartphones"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metadados da categoria"
                    },
                    "500": {
                        "description": "Erro ao listar metadados"
                    }
                }
            }
        },
        "/api/order": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Listar pedidos",
                "description": "Retorna todos os pedidos do usuário autenticado",
                "operationId": "2e0a787dfc4b1239185f92b95e14cbb9",
                "responses": {
                    "200": {
                        "description": "Lista de pedidos"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "500": {
                        "description": "Erro ao buscar pedidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/order/{id}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Detalhes do pedido",
                "description": "Retorna informações completas de um pedido específico",
                "operationId": "1003deedfe383a0f9d1479021de16297",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do pedido"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/order/{id}/session": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Obter sessão de pagamento Stripe",
                "description": "Retorna informações da sessão de pagamento",
                "operationId": "6fe9c16d70d7c7dfc2ff57231d0a939c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sessão criada"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticação via Laravel Sanctum. Use: Bearer {seu-token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}